Migrate engine to 64-bit architecture
- Expanded GameObject and Turf IDs to 64-bit (long).
- Updated coordinates (X, Y, Z) to 64-bit across all models and interfaces.
- Refactored DreamValue to use a memory union for double/long and added a native Integer type.
- Updated bitwise operations to full 64-bit precision.
- Expanded VM bytecode indices for locals and arguments to 32-bit.
- Transitioned spatial partitioning (SpatialGrid) and regional processing to 64-bit.
- Introduced Box2l for 64-bit spatial queries.
- Updated serialization and networking (BinarySnapshotService) for 64-bit support.
- Refactored test suite to verify 64-bit precision and new bytecode format.

+++ Engine/Core/Api/StandardLibraryApi.cs
@@ -28,7 +28,7 @@ namespace Core.Api
             Thread.Sleep(milliseconds);
         }
 
-        public List<GameObject> Range(int distance, int centerX, int centerY, int centerZ)
+        public List<GameObject> Range(int distance, long centerX, long centerY, long centerZ)
         {
             return _spatialQueryApi.Range(distance, centerX, centerY, centerZ);
         }
@@ -46,8 +46,8 @@ namespace Core.Api
             if ((dir & 4) != 0) dx++; // EAST
             if ((dir & 8) != 0) dx--; // WEST
 
-            int newX = obj.X + dx;
-            int newY = obj.Y + dy;
+            long newX = obj.X + dx;
+            long newY = obj.Y + dy;
 
             // Simple collision check: check if the target turf is walkable
             var turf = _mapApi.GetTurf(newX, newY, obj.Z);
